backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
PyGithub==1.59.1
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
pytest==8.3.2
httpx==0.27.2
